+++ Controls-default-theme/_aliases.less
@@ -125,6 +125,7 @@
 //ОБВОДКА КОНТРОЛОВ
 //Толщина обводки контролов.
 @border-thickness:                                                          1px;
+@border-thickness-m:                                                        3px;
 
 //ШРИФТЫ
 //Шрифт по-умолчанию.
