+++ Controls-default-theme/_notAliasesToDelete.less
@@ -285,6 +285,7 @@
 @background-color_Grid-row_hover_val:                                     #F0F5FB;
 @background-color_Grid-row_hover-transparent_val:                         rgba(240,245,251,0);
 @border-color_Grid-row-separator_val:                                     #F5F5F5;
+@border-color_Grid-row-separator-m_val:                                   #F3F3F3;
 @spacing_Grid-editArrow_left_val:                                         8px;
 // размер svg-иконки для сортировки
 @height_SortingArrow-icon_val:                                            12px;
