+++ Controls-default-theme/controls/_Grid.less
@@ -19,11 +19,13 @@
 @border-color_Grid-firstRow-separator:                  @color_separator;
 @border-color_Grid-lastRow-separator:                   @color_separator;
 @border-color_Grid-row-separator:                       @border-color_Grid-row-separator_val;
+@border-color_Grid-row-separator-m:                     @border-color_Grid-row-separator-m_val;
 @height_Grid-results-cell:                              @height_Grid-results-cell_val;
 @min-height_Grid-row:                                   @min-height_Grid-row_val;
 @min-height_Grid-header:                                @height_control_default;
 @min-height_Grid-multi-header:                          @height_control_multi_header;
 @height_Grid-row-separator:                             @thickness_border_default;
+@height_Grid-row-separator-m:                           @border-thickness-m;
 @width_Grid-selection-marker:                           @width-marker_m;
 @color_Grid-selection-marker:                           @marker_color;
 @width_Grid-cell-checkbox:                              @click-area_width_checkbox;
