+++ Controls-demo/grid/RowSeparator/Base/Base.wml
@@ -1,0 +1,8 @@
+<div class="controlsDemo__wrapper">
+    <Controls.grid:View
+            keyProperty="id"
+            attr:class="controlsDemo__inline-flex"
+            source="{{_viewSource}}"
+            rowSeparatorVisibility="{{true}}"
+            columns="{{_columns}}"/>
+</div>
