+++ Controls-demo/grid/RowSeparator/RowSeparator.wml
@@ -1,8 +1,20 @@
 <div class="controlsDemo__wrapper">
-    <Controls.grid:View
-            keyProperty="id"
-            attr:class="controlsDemo__inline-flex"
-            source="{{_viewSource}}"
-            rowSeparatorVisibility="{{true}}"
-            columns="{{_columns}}"/>
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/grid/RowSeparator/Base/Index' }}">
+            <a href="{{content.href}}">Межстрочный разделитель default</a>
+        </Router.router:Reference>
+        <Controls-demo.grid.RowSeparator.Base.Index/>
+    </div>
+
+    <div class="controlsDemo__cell">
+        <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
+                                 state="/Controls-demo/app/:app"
+                                 app="{{ 'Controls-demo/grid/RowSeparator/Wide/Index' }}">
+            <a href="{{content.href}}">Межстрочный разделитель широкий</a>
+        </Router.router:Reference>
+        <Controls-demo.grid.RowSeparator.Wide.Index/>
+    </div>
+
 </div>
