+++ Controls-demo/grid/RowSeparator/Wide/Wide.wml
@@ -1,0 +1,9 @@
+<div class="controlsDemo__wrapper">
+    <Controls.grid:View
+            keyProperty="id"
+            attr:class="controlsDemo__inline-flex"
+            source="{{_viewSource}}"
+            rowSeparatorVisibility="{{true}}"
+            rowSeparatorSize="m"
+            columns="{{_columns}}"/>
+</div>
