+++ Controls/_grid/layout/common/_RowSeparator.less
@@ -1,7 +1,14 @@
 // Разделитель задается верхней границей, которая присутствует всегда (чтобы при включении разделителя таблица не скакала).
 // При выключенном разделителе она прозрачная.
-.controls-Grid__row-cell_theme-@{themeName} {
+.controls-Grid__row-cell_rowSeparatorSize-s_theme-@{themeName} {
    border-top-width: @height_Grid-row-separator;
+}
+
+.controls-Grid__row-cell_rowSeparatorSize-m_theme-@{themeName} {
+   border-top-width: @border-thickness-m;
+}
+
+.controls-Grid__row-cell_theme-@{themeName} {
    border-top-style: solid;
 }
 
@@ -16,6 +23,11 @@
    min-height: @min-height_Grid-row;
 }
 
+.controls-Grid__row-cell_withRowSeparator-m_theme-@{themeName} {
+   border-top-color: @border-color_Grid-row-separator-m;
+   min-height: @min-height_Grid-row;
+}
+
 // Первая строка в таблице с включенным разделителем
 .controls-GridViewV__itemsContainer .controls-Grid__row:first-child .controls-Grid__row-cell_withRowSeparator_theme-@{themeName} {
    border-top-color: @border-color_Grid-firstRow-separator;
@@ -26,6 +38,18 @@
    border-bottom-style: solid;
    border-bottom-color: @border-color_Grid-lastRow-separator;
    min-height: @min-height_Grid-row + @height_Grid-row-separator;
+}
+
+// Первая строка в таблице с включенным разделителем
+.controls-GridViewV__itemsContainer .controls-Grid__row:first-child .controls-Grid__row-cell_withRowSeparator-m_theme-@{themeName} {
+   border-top-color: @border-color_Grid-row-separator-m;
+}
+// Последняя строка в таблице с включенным разделителем
+.controls-GridViewV__itemsContainer .controls-Grid__row:last-child .controls-Grid__row-cell_withRowSeparator-m_theme-@{themeName} {
+   border-bottom-width: @border-thickness-m;
+   border-bottom-style: solid;
+   border-bottom-color: @border-color_Grid-row-separator-m;
+   min-height: @min-height_Grid-row + @border-thickness-m;
 }
